+++ src/src_core/fmg/core/mosaic/cells/Rhombus1.cs
@@ -213,7 +213,7 @@ namespace fmg.core.mosaic.cells {
          }
       }
 
-      public override RectDouble getRcInner(int borderWidth) {
+      public override RectDouble getRcInner(double borderWidth) {
          var attr = Attr;
          var a = attr.A;
          var c = attr.C;
