+++ src/src_core/fmg/core/mosaic/cells/Square1.cs
@@ -95,7 +95,7 @@ namespace fmg.core.mosaic.cells {
          region.SetPoint(3, x1, y1);
       }
 
-      public override RectDouble getRcInner(int borderWidth) {
+      public override RectDouble getRcInner(double borderWidth) {
          var attr = Attr;
          var sq = attr.GetSq(borderWidth);
          var w = borderWidth/2.0;
