+++ src/src_core/fmg/core/mosaic/cells/Trapezoid3.cs
@@ -417,7 +417,7 @@ namespace fmg.core.mosaic.cells {
          }
       }
 
-      public override RectDouble getRcInner(int borderWidth) {
+      public override RectDouble getRcInner(double borderWidth) {
          var attr = Attr;
          var a = attr.A;
          var b = attr.B;
