+++ src/src_core/fmg/core/mosaic/cells/Triangle4.cs
@@ -1059,7 +1059,7 @@ namespace fmg.core.mosaic.cells {
          }
       }
 
-      public override RectDouble getRcInner(int borderWidth) {
+      public override RectDouble getRcInner(double borderWidth) {
          var attr = Attr;
          var w = borderWidth/2.0;
          var sq    = attr.GetSq(borderWidth);
